assign works efficiently (how it should), plus a few other efficiency tweaks

+++ bp_vector/util.h
@@ -2,6 +2,7 @@
 #define BP_VECTOR_UTIL
 
 #include <cassert>
+#include <cstdint>
 
 /* debug printing */
 
@@ -39,7 +40,7 @@ inline T implicit_cast(typename detail::icast_identity<T>::type x)
 }
 
 /* from http://stackoverflow.com/a/9194117 */
-constexpr inline int next_multiple(int i, int m)
+constexpr inline int64_t next_multiple(int64_t i, int64_t m)
 {
     assert(m && ((m & (m-1)) == 0));
     return (i + m-1) & ~(m-1);
